[FRONT | API] Add Ability to Filter Objects by Creator (#issue1512)
