+++ opencti-platform/opencti-front/src/private/components/analysis/reports/ReportLine.js
@@ -1,8 +1,6 @@
-import React, { Component } from 'react';
-import * as PropTypes from 'prop-types';
+import React from 'react';
 import { Link } from 'react-router-dom';
 import { graphql, createFragmentContainer } from 'react-relay';
-import withStyles from '@mui/styles/withStyles';
 import ListItem from '@mui/material/ListItem';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
@@ -10,16 +8,16 @@ import {
   KeyboardArrowRightOutlined,
   DescriptionOutlined,
 } from '@mui/icons-material';
-import { compose, pathOr } from 'ramda';
 import Checkbox from '@mui/material/Checkbox';
 import Skeleton from '@mui/material/Skeleton';
 import Tooltip from '@mui/material/Tooltip';
-import inject18n from '../../../../components/i18n';
+import makeStyles from '@mui/styles/makeStyles';
+import { useFormatter } from '../../../../components/i18n';
 import ItemStatus from '../../../../components/ItemStatus';
 import StixCoreObjectLabels from '../../common/stix_core_objects/StixCoreObjectLabels';
 import ItemMarkings from '../../../../components/ItemMarkings';
 
-const styles = (theme) => ({
+const useStyles = makeStyles((theme) => ({
   item: {
     paddingLeft: 10,
     height: 50,
@@ -48,133 +46,120 @@ const styles = (theme) => ({
     height: '1em',
     backgroundColor: theme.palette.grey[700],
   },
-});
+}));
 
-class ReportLineComponent extends Component {
-  render() {
-    const {
-      fd,
-      classes,
-      node,
-      dataColumns,
-      onLabelClick,
-      onToggleEntity,
-      selectedElements,
-      deSelectedElements,
-      selectAll,
-      onToggleShiftEntity,
-      index,
-    } = this.props;
-    return (
-      <ListItem
-        classes={{ root: classes.item }}
-        divider={true}
-        button={true}
-        component={Link}
-        to={`/dashboard/analysis/reports/${node.id}`}
+const ReportLineComponent = (props) => {
+  const classes = useStyles();
+  const { fd } = useFormatter();
+  const {
+    node,
+    dataColumns,
+    onLabelClick,
+    onToggleEntity,
+    selectedElements,
+    deSelectedElements,
+    selectAll,
+    onToggleShiftEntity,
+    index,
+  } = props;
+  return (
+    <ListItem
+      classes={{ root: classes.item }}
+      divider={true}
+      button={true}
+      component={Link}
+      to={`/dashboard/analysis/reports/${node.id}`}
+    >
+      <ListItemIcon
+        classes={{ root: classes.itemIcon }}
+        style={{ minWidth: 40 }}
+        onClick={(event) => (event.shiftKey
+          ? onToggleShiftEntity(index, node, event)
+          : onToggleEntity(node, event))
+        }
       >
-        <ListItemIcon
-          classes={{ root: classes.itemIcon }}
-          style={{ minWidth: 40 }}
-          onClick={(event) => (event.shiftKey
-            ? onToggleShiftEntity(index, node, event)
-            : onToggleEntity(node, event))
+        <Checkbox
+          edge="start"
+          checked={
+            (selectAll && !(node.id in (deSelectedElements || {})))
+            || node.id in (selectedElements || {})
           }
-        >
-          <Checkbox
-            edge="start"
-            checked={
-              (selectAll && !(node.id in (deSelectedElements || {})))
-              || node.id in (selectedElements || {})
-            }
-            disableRipple={true}
-          />
-        </ListItemIcon>
-        <ListItemIcon classes={{ root: classes.itemIcon }}>
-          <DescriptionOutlined />
-        </ListItemIcon>
-        <ListItemText
-          primary={
-            <div>
-              <Tooltip title={node.name}>
-                <div
-                  className={classes.bodyItem}
-                  style={{ width: dataColumns.name.width }}
-                >
-                  {node.name}
-                </div>
-              </Tooltip>
-              <div
-                className={classes.bodyItem}
-                style={{ width: dataColumns.createdBy.width }}
-              >
-                {pathOr('', ['createdBy', 'name'], node)}
-              </div>
-              <div
-                className={classes.bodyItem}
-                style={{ width: dataColumns.objectLabel.width }}
-              >
-                <StixCoreObjectLabels
-                  variant="inList"
-                  labels={node.objectLabel}
-                  onClick={onLabelClick.bind(this)}
-                />
-              </div>
-              <div
-                className={classes.bodyItem}
-                style={{ width: dataColumns.published.width }}
-              >
-                {fd(node.published)}
-              </div>
-              <div
-                className={classes.bodyItem}
-                style={{ width: dataColumns.x_opencti_workflow_id.width }}
-              >
-                <ItemStatus
-                  status={node.status}
-                  variant="inList"
-                  disabled={!node.workflowEnabled}
-                />
-              </div>
-              <div
-                className={classes.bodyItem}
-                style={{ width: dataColumns.objectMarking.width }}
-              >
-                <ItemMarkings
-                  markingDefinitions={pathOr(
-                    [],
-                    ['objectMarking', 'edges'],
-                    node,
-                  )}
-                  limit={1}
-                  variant="inList"
-                />
-              </div>
-            </div>
-          }
+          disableRipple={true}
         />
-        <ListItemIcon classes={{ root: classes.goIcon }}>
-          <KeyboardArrowRightOutlined />
-        </ListItemIcon>
-      </ListItem>
-    );
-  }
-}
-
-ReportLineComponent.propTypes = {
-  dataColumns: PropTypes.object,
-  node: PropTypes.object,
-  classes: PropTypes.object,
-  fd: PropTypes.func,
-  t: PropTypes.func,
-  onLabelClick: PropTypes.func,
-  onToggleEntity: PropTypes.func,
-  selectedElements: PropTypes.object,
-  deSelectedElements: PropTypes.object,
-  selectAll: PropTypes.bool,
+      </ListItemIcon>
+      <ListItemIcon classes={{ root: classes.itemIcon }}>
+        <DescriptionOutlined />
+      </ListItemIcon>
+      <ListItemText
+        primary={
+          <div>
+            <Tooltip title={node.name}>
+              <div
+                className={classes.bodyItem}
+                style={{ width: dataColumns.name.width }}
+              >
+                {node.name}
+              </div>
+            </Tooltip>
+            <div
+              className={classes.bodyItem}
+              style={{ width: dataColumns.createdBy.width }}
+            >
+              {node.createdBy?.name}
+            </div>
+            <div
+              className={classes.bodyItem}
+              style={{ width: dataColumns.creator.width }}
+            >
+              {node.creator?.name}
+            </div>
+            <div
+              className={classes.bodyItem}
+              style={{ width: dataColumns.objectLabel.width }}
+            >
+              <StixCoreObjectLabels
+                variant="inList"
+                labels={node.objectLabel}
+                onClick={onLabelClick}
+              />
+            </div>
+            <div
+              className={classes.bodyItem}
+              style={{ width: dataColumns.published.width }}
+            >
+              {fd(node.published)}
+            </div>
+            <div
+              className={classes.bodyItem}
+              style={{ width: dataColumns.x_opencti_workflow_id.width }}
+            >
+              <ItemStatus
+                status={node.status}
+                variant="inList"
+                disabled={!node.workflowEnabled}
+              />
+            </div>
+            <div
+              className={classes.bodyItem}
+              style={{ width: dataColumns.objectMarking.width }}
+            >
+              <ItemMarkings
+                markingDefinitions={node.objectMarking.edges ?? []}
+                limit={1}
+                variant="inList"
+              />
+            </div>
+          </div>
+        }
+      />
+      <ListItemIcon classes={{ root: classes.goIcon }}>
+        <KeyboardArrowRightOutlined />
+      </ListItemIcon>
+    </ListItem>
+  );
 };
 
-const ReportLineFragment = createFragmentContainer(ReportLineComponent, {
+export const ReportLine = createFragmentContainer(ReportLineComponent, {
   node: graphql`
     fragment ReportLine_node on Report {
       id
@@ -207,6 +192,10 @@ const ReportLineFragment = createFragmentContainer(ReportLineComponent, {
           }
         }
       }
+      creator {
+        id
+        name
+      }
       status {
         id
         order
@@ -220,116 +209,111 @@ const ReportLineFragment = createFragmentContainer(ReportLineComponent, {
   `,
 });
 
-export const ReportLine = compose(
-  inject18n,
-  withStyles(styles),
-)(ReportLineFragment);
-
-class ReportLineDummyComponent extends Component {
-  render() {
-    const { classes, dataColumns } = this.props;
-    return (
-      <ListItem classes={{ root: classes.item }} divider={true}>
-        <ListItemIcon
-          classes={{ root: classes.itemIconDisabled }}
-          style={{ minWidth: 40 }}
-        >
-          <Checkbox edge="start" disabled={true} disableRipple={true} />
-        </ListItemIcon>
-        <ListItemIcon classes={{ root: classes.itemIcon }}>
-          <Skeleton
-            animation="wave"
-            variant="circular"
-            width={30}
-            height={30}
-          />
-        </ListItemIcon>
-        <ListItemText
-          primary={
-            <div>
-              <div
-                className={classes.bodyItem}
-                style={{ width: dataColumns.name.width }}
-              >
-                <Skeleton
-                  animation="wave"
-                  variant="rectangular"
-                  width="90%"
-                  height="100%"
-                />
-              </div>
-              <div
-                className={classes.bodyItem}
-                style={{ width: dataColumns.createdBy.width }}
-              >
-                <Skeleton
-                  animation="wave"
-                  variant="rectangular"
-                  width="90%"
-                  height="100%"
-                />
-              </div>
-              <div
-                className={classes.bodyItem}
-                style={{ width: dataColumns.objectLabel.width }}
-              >
-                <Skeleton
-                  animation="wave"
-                  variant="rectangular"
-                  width="90%"
-                  height="100%"
-                />
-              </div>
-              <div
-                className={classes.bodyItem}
-                style={{ width: dataColumns.published.width }}
-              >
-                <Skeleton
-                  animation="wave"
-                  variant="rectangular"
-                  width="90%"
-                  height="100%"
-                />
-              </div>
-              <div
-                className={classes.bodyItem}
-                style={{ width: dataColumns.x_opencti_workflow_id.width }}
-              >
-                <Skeleton
-                  animation="wave"
-                  variant="rectangular"
-                  width="90%"
-                  height="100%"
-                />
-              </div>
-              <div
-                className={classes.bodyItem}
-                style={{ width: dataColumns.objectMarking.width }}
-              >
-                <Skeleton
-                  animation="wave"
-                  variant="rectangular"
-                  width={100}
-                  height="100%"
-                />
-              </div>
-            </div>
-          }
+export const ReportLineDummy = (props) => {
+  const classes = useStyles();
+  const { dataColumns } = props;
+  return (
+    <ListItem classes={{ root: classes.item }} divider={true}>
+      <ListItemIcon
+        classes={{ root: classes.itemIconDisabled }}
+        style={{ minWidth: 40 }}
+      >
+        <Checkbox edge="start" disabled={true} disableRipple={true} />
+      </ListItemIcon>
+      <ListItemIcon classes={{ root: classes.itemIcon }}>
+        <Skeleton
+          animation="wave"
+          variant="circular"
+          width={30}
+          height={30}
         />
-        <ListItemIcon classes={{ root: classes.goIcon }}>
-          <KeyboardArrowRightOutlined />
-        </ListItemIcon>
-      </ListItem>
-    );
-  }
-}
-
-ReportLineDummyComponent.propTypes = {
-  classes: PropTypes.object,
-  dataColumns: PropTypes.object,
+      </ListItemIcon>
+      <ListItemText
+        primary={
+          <div>
+            <div
+              className={classes.bodyItem}
+              style={{ width: dataColumns.name.width }}
+            >
+              <Skeleton
+                animation="wave"
+                variant="rectangular"
+                width="90%"
+                height="100%"
+              />
+            </div>
+            <div
+              className={classes.bodyItem}
+              style={{ width: dataColumns.createdBy.width }}
+            >
+              <Skeleton
+                animation="wave"
+                variant="rectangular"
+                width="90%"
+                height="100%"
+              />
+            </div>
+            <div
+              className={classes.bodyItem}
+              style={{ width: dataColumns.creator.width }}
+            >
+              <Skeleton
+                animation="wave"
+                variant="rectangular"
+                width="90%"
+                height="100%"
+              />
+            </div>
+            <div
+              className={classes.bodyItem}
+              style={{ width: dataColumns.objectLabel.width }}
+            >
+              <Skeleton
+                animation="wave"
+                variant="rectangular"
+                width="90%"
+                height="100%"
+              />
+            </div>
+            <div
+              className={classes.bodyItem}
+              style={{ width: dataColumns.published.width }}
+            >
+              <Skeleton
+                animation="wave"
+                variant="rectangular"
+                width="90%"
+                height="100%"
+              />
+            </div>
+            <div
+              className={classes.bodyItem}
+              style={{ width: dataColumns.x_opencti_workflow_id.width }}
+            >
+              <Skeleton
+                animation="wave"
+                variant="rectangular"
+                width="90%"
+                height="100%"
+              />
+            </div>
+            <div
+              className={classes.bodyItem}
+              style={{ width: dataColumns.objectMarking.width }}
+            >
+              <Skeleton
+                animation="wave"
+                variant="rectangular"
+                width={100}
+                height="100%"
+              />
+            </div>
+          </div>
+        }
+      />
+      <ListItemIcon classes={{ root: classes.goIcon }}>
+        <KeyboardArrowRightOutlined />
+      </ListItemIcon>
+    </ListItem>
+  );
 };
-
-export const ReportLineDummy = compose(
-  inject18n,
-  withStyles(styles),
-)(ReportLineDummyComponent);
