+++ opencti-platform/opencti-front/src/private/components/arsenal/vulnerabilities/VulnerabilitiesLines.tsx
@@ -1,0 +1,121 @@
+import React, { FunctionComponent } from 'react';
+import { graphql, PreloadedQuery } from 'react-relay';
+import ListLinesContent from '../../../../components/list_lines/ListLinesContent';
+import { VulnerabilityLine, VulnerabilityLineDummy } from './VulnerabilityLine';
+import { DataColumns } from '../../../../components/list_lines';
+import { UseLocalStorage } from '../../../../utils/hooks/useLocalStorage';
+import {
+  VulnerabilitiesLinesPaginationQuery,
+  VulnerabilitiesLinesPaginationQuery$variables,
+} from './__generated__/VulnerabilitiesLinesPaginationQuery.graphql';
+import usePreloadedPaginationFragment from '../../../../utils/hooks/usePreloadedPaginationFragment';
+import { VulnerabilitiesLines_data$key } from './__generated__/VulnerabilitiesLines_data.graphql';
+
+const nbOfRowsToLoad = 50;
+
+interface VulnerabilityLinesProps {
+  queryRef: PreloadedQuery<VulnerabilitiesLinesPaginationQuery>,
+  dataColumns: DataColumns,
+  paginationOptions?: VulnerabilitiesLinesPaginationQuery$variables,
+  setNumberOfElements: UseLocalStorage[2]['handleSetNumberOfElements'],
+  onLabelClick: (k: string, id: string, value: Record<string, unknown>, event: React.KeyboardEvent) => void,
+}
+
+export const vulnerabilitiesLinesQuery = graphql`
+  query VulnerabilitiesLinesPaginationQuery(
+    $search: String
+    $count: Int!
+    $cursor: ID
+    $orderBy: VulnerabilitiesOrdering
+    $orderMode: OrderingMode
+    $filters: [VulnerabilitiesFiltering!]
+  ) {
+    ...VulnerabilitiesLines_data
+    @arguments(
+      search: $search
+      count: $count
+      cursor: $cursor
+      orderBy: $orderBy
+      orderMode: $orderMode
+      filters: $filters
+    )
+  }
+`;
+
+const vulnerabilitiesLinesFragment = graphql`
+  fragment VulnerabilitiesLines_data on Query
+  @argumentDefinitions(
+    search: { type: "String" }
+    count: { type: "Int", defaultValue: 25 }
+    cursor: { type: "ID" }
+    orderBy: { type: "VulnerabilitiesOrdering", defaultValue: name }
+    orderMode: { type: "OrderingMode", defaultValue: asc }
+    filters: { type: "[VulnerabilitiesFiltering!]" }
+  ) @refetchable(queryName: "VulnerabilitiesLinesRefetchQuery") {
+    vulnerabilities(
+      search: $search
+      first: $count
+      after: $cursor
+      orderBy: $orderBy
+      orderMode: $orderMode
+      filters: $filters
+    ) @connection(key: "Pagination_vulnerabilities") {
+      edges {
+        node {
+          id
+          name
+          description
+          creator {
+            id
+            name
+          }
+          ...VulnerabilityLine_node
+        }
+      }
+      pageInfo {
+        endCursor
+        hasNextPage
+        globalCount
+      }
+    }
+  }
+`;
+
+const VulnerabilitiesLines: FunctionComponent<VulnerabilityLinesProps> = ({
+  setNumberOfElements,
+  queryRef,
+  dataColumns,
+  paginationOptions,
+  onLabelClick }) => {
+  const {
+    data,
+    hasMore,
+    loadMore,
+    isLoadingMore,
+  } = usePreloadedPaginationFragment<VulnerabilitiesLinesPaginationQuery, VulnerabilitiesLines_data$key>({
+    linesQuery: vulnerabilitiesLinesQuery,
+    linesFragment: vulnerabilitiesLinesFragment,
+    queryRef,
+    nodePath: ['vulnerabilities', 'edges'],
+    setNumberOfElements,
+  });
+
+  return (
+    <ListLinesContent
+      initialLoading={!data}
+      loadMore={loadMore}
+      hasMore={hasMore}
+      isLoading={isLoadingMore}
+      dataList={data?.vulnerabilities?.edges ?? []}
+      globalCount={data?.vulnerabilities?.pageInfo?.globalCount ?? nbOfRowsToLoad}
+      LineComponent={VulnerabilityLine}
+      DummyLineComponent={VulnerabilityLineDummy}
+      dataColumns={dataColumns}
+      nbOfRowsToLoad={nbOfRowsToLoad}
+      paginationOptions={paginationOptions}
+      onLabelClick={onLabelClick}
+    />
+  );
+};
+
+export default VulnerabilitiesLines;
