+++ opencti-platform/opencti-front/src/private/components/arsenal/vulnerabilities/VulnerabilityLine.tsx
@@ -1,20 +1,22 @@
-import React from 'react';
+import React, { FunctionComponent } from 'react';
 import { Link } from 'react-router-dom';
-import { graphql, createFragmentContainer } from 'react-relay';
+import { graphql, useFragment } from 'react-relay';
 import ListItem from '@mui/material/ListItem';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
-import { KeyboardArrowRight } from '@mui/icons-material';
-import { HexagonOutline } from 'mdi-material-ui';
-import Checkbox from '@mui/material/Checkbox';
+import {
+  KeyboardArrowRightOutlined,
+  BugReportOutlined,
+} from '@mui/icons-material';
 import Skeleton from '@mui/material/Skeleton';
 import makeStyles from '@mui/styles/makeStyles';
 import { useFormatter } from '../../../../components/i18n';
 import StixCoreObjectLabels from '../../common/stix_core_objects/StixCoreObjectLabels';
-import ItemMarkings from '../../../../components/ItemMarkings';
-import { renderObservableValue } from '../../../../utils/String';
+import { Theme } from '../../../../components/Theme';
+import { VulnerabilityLine_node$key } from './__generated__/VulnerabilityLine_node.graphql';
+import { DataColumns } from '../../../../components/list_lines';
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles<Theme>((theme) => ({
   item: {
     paddingLeft: 10,
     height: 50,
@@ -36,72 +38,78 @@ const useStyles = makeStyles((theme) => ({
     right: -10,
   },
   itemIconDisabled: {
-    color: theme.palette.grey[700],
+    color: theme.palette.grey?.[700],
   },
   placeholder: {
     display: 'inline-block',
     height: '1em',
-    backgroundColor: theme.palette.grey[700],
+    backgroundColor: theme.palette.grey?.[700],
   },
 }));
 
-const StixCyberObservableLineComponent = ({
+interface VulnerabilityLineProps {
+  node: VulnerabilityLine_node$key,
+  dataColumns: DataColumns,
+  onLabelClick: (k: string, id: string, value: Record<string, unknown>, event: React.KeyboardEvent) => void,
+}
+
+const vulnerabilityLineFragment = graphql`
+  fragment VulnerabilityLine_node on Vulnerability {
+    id
+    name
+    x_opencti_base_severity
+    created
+    modified
+    objectLabel {
+      edges {
+        node {
+          id
+          value
+          color
+        }
+      }
+    }
+    creator {
+      id
+      name
+    }
+  }
+`;
+
+export const VulnerabilityLine: FunctionComponent<VulnerabilityLineProps> = ({
   dataColumns,
   node,
   onLabelClick,
-  onToggleEntity,
-  selectedElements,
-  deSelectedElements,
-  selectAll,
-  onToggleShiftEntity,
-  index,
 }) => {
   const classes = useStyles();
-  const { t, nsdt } = useFormatter();
+  const { fd } = useFormatter();
+
+  const data = useFragment(vulnerabilityLineFragment, node);
   return (
     <ListItem
       classes={{ root: classes.item }}
       divider={true}
       button={true}
       component={Link}
-      to={`/dashboard/observations/${
-        node.entity_type === 'Artifact' ? 'artifacts' : 'observables'
-      }/${node.id}`}
+      to={`/dashboard/arsenal/vulnerabilities/${data.id}`}
     >
-      <ListItemIcon
-        classes={{ root: classes.itemIcon }}
-        style={{ minWidth: 40 }}
-        onClick={(event) => (event.shiftKey
-          ? onToggleShiftEntity(index, node, event)
-          : onToggleEntity(node, event))
-        }
-      >
-        <Checkbox
-          edge="start"
-          checked={
-            (selectAll && !(node.id in (deSelectedElements || {})))
-            || node.id in (selectedElements || {})
-          }
-          disableRipple={true}
-        />
-      </ListItemIcon>
       <ListItemIcon classes={{ root: classes.itemIcon }}>
-        <HexagonOutline />
+        <BugReportOutlined />
       </ListItemIcon>
       <ListItemText
         primary={
           <div>
             <div
               className={classes.bodyItem}
-              style={{ width: dataColumns.entity_type.width }}
+              style={{ width: dataColumns.name.width }}
             >
-              {t(`entity_${node.entity_type}`)}
+              {data.name}
             </div>
             <div
               className={classes.bodyItem}
-              style={{ width: dataColumns.observable_value.width }}
+              style={{ width: dataColumns.x_opencti_base_severity.width }}
             >
-              {renderObservableValue(node)}
+              {data.x_opencti_base_severity}
             </div>
             <div
               className={classes.bodyItem}
@@ -109,119 +117,56 @@ const StixCyberObservableLineComponent = ({
             >
               <StixCoreObjectLabels
                 variant="inList"
-                labels={node.objectLabel}
+                labels={data.objectLabel}
                 onClick={onLabelClick}
               />
             </div>
             <div
               className={classes.bodyItem}
-              style={{ width: dataColumns.created_at.width }}
+              style={{ width: dataColumns.created.width }}
             >
-              {nsdt(node.created_at)}
+              {fd(data.created)}
+            </div>
+            <div
+              className={classes.bodyItem}
+              style={{ width: dataColumns.modified.width }}
+            >
+              {fd(data.modified)}
             </div>
             <div
               className={classes.bodyItem}
               style={{ width: dataColumns.creator.width }}
             >
-              {t(node.creator.name)}
-            </div>
-            <div
-              className={classes.bodyItem}
-              style={{ width: dataColumns.objectMarking.width }}
-            >
-              <ItemMarkings
-                markingDefinitions={node.objectMarking.edges ?? []}
-                limit={1}
-                variant="inList"
-              />
+              {data.creator?.name}
             </div>
           </div>
         }
       />
       <ListItemIcon classes={{ root: classes.goIcon }}>
-        <KeyboardArrowRight />
+        <KeyboardArrowRightOutlined />
       </ListItemIcon>
     </ListItem>
   );
 };
 
-export const StixCyberObservableLine = createFragmentContainer(
-  StixCyberObservableLineComponent,
-  {
-    node: graphql`
-      fragment StixCyberObservableLine_node on StixCyberObservable {
-        id
-        entity_type
-        parent_types
-        observable_value
-        created_at
-        ... on IPv4Addr {
-          countries {
-            edges {
-              node {
-                name
-                x_opencti_aliases
-              }
-            }
-          }
-        }
-        ... on IPv6Addr {
-          countries {
-            edges {
-              node {
-                name
-                x_opencti_aliases
-              }
-            }
-          }
-        }
-        objectMarking {
-          edges {
-            node {
-              id
-              definition
-              x_opencti_color
-            }
-          }
-        }
-        objectLabel {
-          edges {
-            node {
-              id
-              value
-              color
-            }
-          }
-        }
-        creator {
-          id
-          name
-        }
-      }
-    `,
-  },
-);
-
-export const StixCyberObservableLineDummy = (props) => {
+export const VulnerabilityLineDummy = ({ dataColumns }: { dataColumns: DataColumns }) => {
   const classes = useStyles();
-  const { dataColumns } = props;
   return (
     <ListItem classes={{ root: classes.item }} divider={true}>
-      <ListItemIcon
-        classes={{ root: classes.itemIconDisabled }}
-        style={{ minWidth: 40 }}
-      >
-        <Checkbox edge="start" disabled={true} disableRipple={true} />
-      </ListItemIcon>
-      <ListItemIcon classes={{ root: classes.itemIcon }}>
-        <Skeleton animation="wave" variant="circular" width={30} height={30} />
+      <ListItemIcon classes={{ root: classes.itemIconDisabled }}>
+        <Skeleton
+          animation="wave"
+          variant="circular"
+          width={30}
+          height={30}
+        />
       </ListItemIcon>
       <ListItemText
         primary={
           <div>
             <div
               className={classes.bodyItem}
-              style={{ width: dataColumns.entity_type.width }}
+              style={{ width: dataColumns.name.width }}
             >
               <Skeleton
                 animation="wave"
@@ -232,7 +177,7 @@ export const StixCyberObservableLineDummy = (props) => {
             </div>
             <div
               className={classes.bodyItem}
-              style={{ width: dataColumns.observable_value.width }}
+              style={{ width: dataColumns.x_opencti_base_severity.width }}
             >
               <Skeleton
                 animation="wave"
@@ -254,7 +199,18 @@ export const StixCyberObservableLineDummy = (props) => {
             </div>
             <div
               className={classes.bodyItem}
-              style={{ width: dataColumns.created_at.width }}
+              style={{ width: dataColumns.created.width }}
+            >
+              <Skeleton
+                animation="wave"
+                variant="rectangular"
+                width={140}
+                height="100%"
+              />
+            </div>
+            <div
+              className={classes.bodyItem}
+              style={{ width: dataColumns.modified.width }}
             >
               <Skeleton
                 animation="wave"
@@ -274,22 +230,11 @@ export const StixCyberObservableLineDummy = (props) => {
                 height="100%"
               />
             </div>
-            <div
-              className={classes.bodyItem}
-              style={{ width: dataColumns.objectMarking.width }}
-            >
-              <Skeleton
-                animation="wave"
-                variant="rectangular"
-                width={100}
-                height="100%"
-              />
-            </div>
           </div>
         }
       />
       <ListItemIcon classes={{ root: classes.goIcon }}>
-        <KeyboardArrowRight />
+        <KeyboardArrowRightOutlined />
       </ListItemIcon>
     </ListItem>
   );
