+++ opencti-platform/opencti-front/src/private/components/observations/stix_cyber_observables/StixCyberObservableCreation.js
@@ -237,6 +237,10 @@ const stixCyberObservableMutation = graphql`
           }
         }
       }
+      creator {
+        id 
+        name
+      }
     }
   }
 `;
