+++ opencti-platform/opencti-front/src/private/components/observations/stix_cyber_observables/StixCyberObservablesLines.js
@@ -186,6 +186,10 @@ export default createPaginationContainer(
               entity_type
               observable_value
               created_at
+              creator {
+                id
+                name
+              }
               objectMarking {
                 edges {
                   node {
