+++ opencti-platform/opencti-graphql/src/resolvers/report.js
@@ -70,7 +70,11 @@ const reportResolvers = {
     createdBy: buildRefRelationKey(RELATION_CREATED_BY),
     markedBy: buildRefRelationKey(RELATION_OBJECT_MARKING),
     labelledBy: buildRefRelationKey(RELATION_OBJECT_LABEL),
-    objectContains: buildRefRelationKey(RELATION_OBJECT, '*')
+    objectContains: buildRefRelationKey(RELATION_OBJECT, '*'),
+    creator: 'creator_id',
+  },
+  ReportsOrdering: {
+    creator: 'creator_id',
   },
   Mutation: {
     reportEdit: (_, { id }, context) => ({
