+++ opencti-platform/opencti-graphql/src/resolvers/vulnerability.js
@@ -22,6 +22,10 @@ const vulnerabilityResolvers = {
     createdBy: buildRefRelationKey(RELATION_CREATED_BY),
     markedBy: buildRefRelationKey(RELATION_OBJECT_MARKING),
     labelledBy: buildRefRelationKey(RELATION_OBJECT_LABEL),
+    creator: 'creator_id',
+  },
+  VulnerabilitiesOrdering: {
+    creator: 'creator_id',
   },
   Vulnerability: {
     softwares: (vulnerability, _, context) => softwaresLoader.load(vulnerability.id, context, context.user),
