+++ opencti-platform/opencti-graphql/src/schema/stixCyberObservable.js
@@ -100,8 +100,11 @@ export const stixCyberObservableOptions = {
     hashes_MD5: 'hashes.MD5',
     hashes_SHA1: 'hashes.SHA-1',
     hashes_SHA256: 'hashes.SHA-256',
+    creator: 'creator_id',
   },
-  StixCyberObservablesOrdering: {}
+  StixCyberObservablesOrdering: {
+    creator: 'creator_id',
+  }
 };
 
 const stixCyberObservableFieldsToBeUpdated = {
